roles/extrepo/dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --strip-extras dev-requirements.in
#
ansible-core==2.13.13
    # via -r dev-requirements.in
black==23.11.0
    # via -r dev-requirements.in
build==1.0.3
    # via pip-tools
cffi==1.16.0
    # via cryptography
click==8.1.7
    # via
    #   black
    #   pip-tools
cryptography==41.0.7
    # via ansible-core
exceptiongroup==1.2.0
    # via pytest
importlib-metadata==6.9.0
    # via build
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via -r dev-requirements.in
jinja2==3.1.2
    # via ansible-core
markupsafe==2.1.3
    # via jinja2
mypy-extensions==1.0.0
    # via black
packaging==23.2
    # via
    #   ansible-core
    #   black
    #   build
    #   pytest
pathspec==0.11.2
    # via black
pip-tools==7.3.0
    # via -r dev-requirements.in
platformdirs==4.0.0
    # via black
pluggy==1.3.0
    # via pytest
pycparser==2.21
    # via cffi
pyproject-hooks==1.0.0
    # via build
pytest==7.4.3
    # via -r dev-requirements.in
pyyaml==6.0.1
    # via ansible-core
resolvelib==0.8.1
    # via ansible-core
tomli==2.0.1
    # via
    #   black
    #   build
    #   pip-tools
    #   pyproject-hooks
    #   pytest
typing-extensions==4.8.0
    # via black
wheel==0.42.0
    # via pip-tools
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
